requirements.txt
fitz==0.0.1.dev2
python_docx==1.2.0